refactoring introduce unit test for service class

+++ TheShop/Contracts/IShopService.cs
@@ -9,7 +9,7 @@ namespace TheShop.Contracts
 {
     public interface IShopService
     {
-        Article OrderArticle(int id, decimal maxExpectedPrice);
+        Article OrderArticle(int id, double maxExpectedPrice);
         void SellArticle(int buyerId, Article article);
         void DispalyArticle(Article article);
     }
